refactor: Simplify Core class in SpiceDb
Significant changes have been made to the Core class in the SpiceDb project. The Start method has been removed and its functionality integrated into the constructor. The Permissions, Schema, Watch, and Experimental properties are now public readonly fields initialized directly in the constructor. Unnecessary methods for relationship updates have been removed. This refactoring simplifies the code structure and improves readability.
