+++ SpiceDb/Models/Consistency.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SpiceDb.Models;
+
+/*
+/// Unused for now
+public class Consistency
+{
+    public bool MinimizeLatency { get; set; }
+    public ZedToken? AtLeastAsFresh { get; set; } = default!;
+    public ZedToken? AtExactSnapshot { get; set; } = default!;
+    public bool FullyConsistent { get; set; }
+}
+*/
